implements variants of Xaiver
They are several different Xaiver variants used and it is not quite
clear if there is a "correct" one. Different variants can be chosen by
the combinations of a uniform or normal distribution with a specific
calculation of the variance. Currently implemented are different
variants from Caffe, the original one by Bengio and Glorot and the one
by K. He et. al 2015 (called msra in Caffe). The current default is the
one used by the Python interface to MXNet (3 / (in+out)).

implement several variants of xaiver

rewrite xaiver to use enums