add initial image argument
This adds a feature to set the initial image to be optimized from the command line. This can be useful if you want to first run the program with a smaller max_long_edge, then use that for the initial image for a longer max_long_edge as this can help with speed optimizations. This may also be useful for general experimentation.