Correct init idx2name on multi-CPU/GPU training v2
This patch is related to #2327 and #2337, which fix a bug of idx2name in multi-GPU environment for Module.
Note, all previous codes which involved lr_mult and wd_mult in multi-GPU environment may suffer from this bug,
for example, finetuning with lr_mult of lower layers fixed to zero on multi-GPU will give incorrect result.