Fixes deps/build.jl on Travis (#142)
The problem is that we need to specify `MSHADOW_LDFLAGS` and
`MSHADOW_CFLAGS`. The way MXNet is currently setup we can't change these
from the terminal. This solves the problem by using `sed` to change
`mshadow/make/mshadow.mk`. But if mshadow would change git could throw
an error on updating. So we remove the changes before an update and add
them afterwards again. This has the drawback that users can't modifty
that file. (But those users should maintain their own build and use
`MXNET_HOME`.)

This also eagerly deletes `usr/lib/libmxnet.so` So that we actually
trigger a rebuild on an update of the pkg.