Revert the reldiff in test_utils Fix the argument of dot and batch_dot as mentioned in #3791 New operators: topk, sort, argsort, arange Move operator to operator/tensor: BlockGrad Add the argument dtype to initialization operators (_zeros, _ones, _arange) Enable gradient for argmax and argmin by adding the gradient node to _zeros.
Some Notes:

The implementation of TopK is not so optimized at this stage. The implementation is to first sort the data and then keep the top k indices/values. Also, the "axis" argument is supported in topk and topk is equivalent to argsort and sort if we set "k<1" and ret_typ = "indices" or "value"

The arange operator has an additional "repeat" argument, which indicates the number of times each element will be repeated in the generation process.