Pip installer for converter: mxnet-coreml-converter.
Runs only on MacOS and python 2.7. Once inside the directory pip_package, user needs
to run:
python setup.py bdist_wheel
twine upload dist/*

Once uploaded it'll look like this:
https://testpypi.python.org/pypi/mxnet-coreml-converter

Also updated the README for converter to reflect this.

Note that we are going with a package per tool for the time being. Please leave feedback if you think it is better to adopt the policy of all the tools in one single package.

Unit tests continue to pass.