Fix missing of CUDA device on non-GPU host
The issue that "no CUDA-capable device is detected" occurs when calling MxNet from Java/Scala on a host without GPU. 

The user scenario is to run MxNet model on CPU-only host with Context input as cpu(0). The default GPU unit shares the same index of zero with the default CPU unit. By default the MXNET_USE_CUDA is enabled and the logic will proceed to the line of "CUDA_CALL". The calling of cudaSetDevice exits with exception since there is GPU found for index of zero. The proposal here is to enable the "CUDA_CALL" only when the device type of Context is not CPU.