Add ‘num_examples’ parameter to generated CaffeDataIter.
CaffeDataIter requires user to provide number of examples in the dataset using the num_examples parameter.
However this information is not available in Caffe prototxt.

This commit lets user add this information to the caffe prototxt as a comment. Caffe will parse it as comment and ignore it. Translator will pick the value and use it while generating CaffeDataIter.

Example usage:
  data_param {
    source: "data/mnist/mnist_train_lmdb"
    #caffe2mxnet num_examples: 60000
    batch_size: 64
    backend: LMDB
  }