Fixed rtc error on Tesla / Jetson TX2
When rtc was run on a Tesla or Pascal device previously the user would
receive an 'imports not found' error.  This was due to the fact that we
were using compute capability 3 by default which was causing modern devices
to fail to load a correct library.  We now will read from the first cuda
capable device on the system  and use it's compute capability instead of
the default.  If the user does specify a gpu architecture explicitly when
calling rtc then we use the explicit value.