[scala] Fix various resource leaks
These leaks were diagnosed with the leak detection added in the previous
commit. This is not an exhaustive clean up but it allows predicting with a
model from Scala at scale (hundreds of millions of comparisons) without a
reported leak, as well as removing the most common errors when training using
the Module code.

Co-authored-by: Andre Tamm <tammtamm@amazon.com>