Use MKL-ML options only if USE_MKLML=1
The Makefile adds flags to CFLAGS and LDFLAGS without checking if
USE_MKLML is set to 1. If a developer wants to use a full installation
of MKL and specifies it using MKLROOT and also specifies USE_MKL2017=1
then this patch makes sure that MKL_ML specific flags are not added.