Prepend additional LDFLAGS instead of appending
Because order matters!

If a developer specifies additional flags using LDFLAGS then they
should be given more precedence to the flags added by the .mk files. As
an example, right now, the Makefile uses pkg-config to fetch all the
flags provided by opencv and ends up overlinking. If a developer passes
'-Wl,--as-needed' to avoid that via LDFLAGS, it ends up being added to
the end and is thus of no value.