Remove finalizers from Scala API (#8887) (#9568)
* [scala] Remove finalizers for leakable resources

Finalizers always run in a separate thread which is not controlled by the user.
Since MXNet cannot be safely accessed from multiple threads, this causes memory
corruption. It is not safe to use finalizers in this way.

This change also adds some leaked object tracing, to allow leaks to be tracked.
It's easy to leak objects and despite warnings in the documentation leaks are
common (even within the Scala API itself). The leak tracing is activated by a
system property, as its runtime cost may be significant.

With the system property unset, the *first* leak of a type will be reported
(without a trace) as a prompt to the developer to investigate.

Co-authored-by: Andre Tamm <tammtamm@amazon.com>

* [scala] Fix various resource leaks

These leaks were diagnosed with the leak detection added in the previous
commit. This is not an exhaustive clean up but it allows predicting with a
model from Scala at scale (hundreds of millions of comparisons) without a
reported leak, as well as removing the most common errors when training using
the Module code.

Co-authored-by: Andre Tamm <tammtamm@amazon.com>