Fix sign bug in spatial transformer interpolation.
It's possible that the computed `data_index` wants to take negative numbers for
cases when the interpolation lookup falls out of the source window. This is bad
on 64-bit machines as the test for hitting the source window still works but
once the the 32-bit unsigned number is added to the 64-bit pointer things break
with a segmentation fault.