Fixed divison by zero bug in DistanceWeightedSampling in gluon example
Sample selection for training is based on a vector of computed probablities that come from distance weights. In the current implementation these weights can become zero when distance to all other neightbors is zero.
Zero weights lead to nan bugs in the model, this commit is supposed to fix it by changing those zero weights not being divided by zero any longer.