Fix Clojure BERT example's context argument (#14843)
* Clojure BERT example: minor code cleanup

* Remove unused requires
* Remove unused vars & function
* Use `io` alias

* Clojure BERT example: whitespace fix

* Clojure BERT example: allow running with GPU

The `infer` function accepts a CPU/GPU context, which the command line
version of this example exposes as a `:cpu`/`:gpu`
keyword. Previously, these options were ignored and the context was
overridden to the default context (CPU). This commit allows
users (both REPL and shell) to pass in a GPU context.