Don't use global logger. Make it specific to each estimator.
gluon.contrib.estimator used a Logger shared between all event handlers that is
obtained via `logging.getLogger('gluon.contrib.estimator.event_handlers')`. This
logger used to be configured every time creating a
gluon.contrib.estimator.LoggingHandler, which is a bug. We can't modify a global
Logger instance whenever the user creates an Estimator and a LoggingHandler.

Instead, this commit separates the LoggingHandler (responsible for logging
metadata during estimator.fit) from the configuration of the Logger.

We expose the Logger as attribute of the Estimator, and configure it to output
to stdout by default. Instructions are given how users can configure the
Estimator.logger to log to a file instead.