USE_JEMALLOC=OFF by default
As of jemalloc 5, jemalloc default build can not be used in libraries that are dlopened. However, libmxnet.so is dlopened by Python (ctypes). To use MXNet with jemalloc 5, users must not to system libjemalloc.so but must rather link to a libjemalloc compiled with special parameters to allow dlopen to work.

jemalloc 5 is distributed as part of Ubuntu 18.10 and higher, as well as Debian Stable. Users on these systems will be unable to compile MXNet with USE_JEMALLOC=ON.