Add simplified HybridBlock.forward without F (#17530)
Users can now implement HybridBlock.forward instead of HybridBlock.hybrid_forward.
HybridBlock.forward has the same signature as Block.forward. For example:

  class MyBlock(mx.gluon.HybridBlock):
      def __init__(self, *, prefix=None, params=None):
          super().__init__(prefix, params)
          with self.name_scope():
              self.dense = mx.gluon.nn.Dense(units=10)
              self.weight = self.params.get('weight', allow_deferred_init=True)
      def infer_shape(self, x):
          self.weight.shape = (x.shape[1], )
      def forward(self, x):
          return self.dense(x) + self.weight.data(x.context) 

Hybridization of HybridBlock.forward is based on a deferred computation mode in
the MXNet backend, which enables recording computation via tracing in the
mxnet.nd and mxnet.np interfaces. The recorded computation can be exported to a
symbolic representation and is used for optimized execution with the CachedOp.

As tracing is based on the imperative APIs, users can access shape information
of the arrays. As x.shape for some array x is a python tuple, any use of that
shape will be a constant in the recorded graph and may limit the recorded graph
to be used with inputs of the same shape only.

As part of the change from hybrid_forward to forward, we also disable support
for parameter shape inference in the MXNet backend in the case of deferred
parameter initialization. Shape inference in the backend was limited and did by
it's very nature not support dynamic shape operators. Instead, users should now
always implement HybridBlock.infer_shape to set the parameter shapes if the
parameter shape was not set during HybridBlock.__init__. See the example above.

An example of the internal deferred compute APIs is:

  a = mx.np.arange(10)
  dc.set_variable(a, mx.sym.var('a').as_np_ndarray())
  with dc.context():
      b = a ** 2
  symbol = dc.get_symbol(b)