CI: Prime docker cache in "Docker cache build" job (#18438)
Cache priming was previously only handled when using the ./ci/build.py or
./ci/docker_cache.py script. The former is used during most CI jobs. The latter
is used for containers not declared in docker-compose.yml during the "Docker
cache build" job. More and more containers are declared in
docker-compose.yml now, but prior to this commit their cache wasn't primed in the
"Docker cache build" job.