Update symbol visibility control
Prior to this commit, linker version scripts were used to control the set of
visible symbols. These scripts included C++ Symbols, that do not have a stable
ABI and should not have been exported. Instead, we annotate the visiblity in the
source code as per the best practice.

Refer to "Deep CMake For Library Authors" CppCon 2019 talk by Craig Scott for
details on the cmake settings.